adding unsigned short support for memory optimization

+++ vsm/model/_cgs_update.pyx
@@ -5,7 +5,7 @@ import cython
 
 
 def cgs_update(int itr, 
-               int [:] corpus,
+               unsigned short [:] corpus,
                double [:,:] word_top,
                double [:] inv_top_sums,
                double [:,:] top_doc,
